Refactor to use Target Hotglue SDK
