--- target_woocommerce/rest.py
@@ -1,107 +1,0 @@
-"""WoocommerceSink target sink class, which handles writing streams."""
-
-from base64 import b64encode
-from datetime import datetime
-from typing import Any, Callable, Dict, List, Optional, cast
-
-import backoff
-import requests
-from singer_sdk.exceptions import FatalAPIError, RetriableAPIError
-from random_user_agent.user_agent import UserAgent
-
-
-class Rest:
-
-    timeout = 300
-    access_token = None
-    user_agents = UserAgent(software_engines="blink", software_names="chrome")
-
-    @property
-    def authenticator(self):
-        user = self.config.get("consumer_key")
-        passwd = self.config.get("consumer_secret")
-        token = b64encode(f"{user}:{passwd}".encode()).decode()
-        return f"Basic {token}"
-
-    @property
-    def http_headers(self) -> dict:
-        """Return the http headers needed."""
-        headers = {}
-        headers["Content-Type"] = "application/json"
-        headers.update({"Authorization": self.authenticator})
-        headers["User-Agent"] = self.user_agents.get_random_user_agent().strip()
-        if "user_agent" in self.config:
-            headers["User-Agent"] = self.config.get("user_agent")
-        return headers
-
-    @backoff.on_exception(
-        backoff.expo,
-        (RetriableAPIError, requests.exceptions.ReadTimeout),
-        max_tries=5,
-        factor=2,
-    )
-    def _request(
-        self, http_method, endpoint, params=None, request_data=None
-    ) -> requests.PreparedRequest:
-        """Prepare a request object."""
-        url = self.url(endpoint)
-        headers = self.http_headers
-
-        response = requests.request(
-            method=http_method,
-            url=url,
-            params=params,
-            headers=headers,
-            json=request_data,
-        )
-        self.validate_response(response)
-        return response
-
-    def request_api(self, http_method, endpoint=None, params=None, request_data=None):
-        """Request records from REST endpoint(s), returning response records."""
-        resp = self._request(http_method, endpoint, params, request_data)
-        return resp
-
-    def validate_response(self, response: requests.Response) -> None:
-        """Validate HTTP response."""
-        if response.status_code in [429] or 500 <= response.status_code < 600:
-            msg = self.response_error_message(response)
-            raise RetriableAPIError(msg, response)
-        elif 400 <= response.status_code < 500:
-            try:
-                msg = response.text
-            except:
-                msg = self.response_error_message(response)
-            raise FatalAPIError(msg)
-
-    def response_error_message(self, response: requests.Response) -> str:
-        """Build error message for invalid http statuses."""
-        if 400 <= response.status_code < 500:
-            error_type = "Client"
-        else:
-            error_type = "Server"
-
-        return (
-            f"{response.status_code} {error_type} Error: "
-            f"{response.reason} for path: {self.endpoint}"
-        )
-
-    @staticmethod
-    def clean_dict_items(dict):
-        return {k: v for k, v in dict.items() if v not in [None, ""]}
-
-    def clean_payload(self, item):
-        item = self.clean_dict_items(item)
-        output = {}
-        for k, v in item.items():
-            if isinstance(v, datetime):
-                dt_str = v.strftime("%Y-%m-%dT%H:%M:%S%z")
-                if len(dt_str) > 20:
-                    output[k] = f"{dt_str[:-2]}:{dt_str[-2:]}"
-                else:
-                    output[k] = dt_str
-            elif isinstance(v, dict):
-                output[k] = self.clean_payload(v)
-            else:
-                output[k] = v
-        return output
